// gui/src/api/gen/types/DatasetRead.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type DatasetRead = {
    /**
     * @type integer
    */
    id: number;
    /**
     * @type integer
    */
    project_id: number;
    /**
     * @type string
    */
    name: string;
    description?: (string | null);
};